+++ StripeFinancialConnections/StripeFinancialConnections/Source/Native/ManualEntry/ManualEntryViewController.swift
@@ -51,10 +51,7 @@ final class ManualEntryViewController: UIViewController {
 
         let footerView = PaneLayoutView.createFooterView(
             primaryButtonConfiguration: PaneLayoutView.ButtonConfiguration(
-                title: STPLocalizedString(
-                    "Submit",
-                    "The submit button for a screen that allows a user to manually enter their bank account information."
-                ),
+                title: String.Localized.submit,
                 accessibilityIdentifier: "manual_entry_continue_button",
                 action: didSelectContinue
             ),