+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/PaymentSheet+PaymentMethodAvailability.swift
@@ -49,6 +49,7 @@ extension PaymentSheet {
         .krCard,
         .payco,
         .sequra,
+        .scalapay,
     ]
 
     /// A list of `STPPaymentMethodType` that can be saved in PaymentSheet